Add support for extensions
If loaded from a helper then you can specify a `resources folder` to be passed in as the `resource path` when calling `load extension`.

If no explicit resource folder has been specified, and a `./resources` folder exists alongside the `app.yml` file, then the `./resources` folder will be specified as the `resource path` folder for the extension.