Path name variables weren’t being processed at the right time
Variables in path names were being replaced at the end of handlers that processed arrays of files. This wouldn’t work if the array entry pointed to a folder as the code could not successfully loop through the folder contents. The variable hadn’t been replaced yet! Now variables are replaced before the array is processed.