Rename "components" to "ui"
"ui" is more descriptive of how these stacks are used in a levure application. The `stacks` property was rename to `ui` for helpers as well. A new `autoload` setting has been added for libraries, frontscripts, and backscripts. A helper can use the setting to specify a library, frontscript, or backscript that should not be inserted when app loads. Previously a helper had to use the `stacks` property to make a stack available in memory but not have it put in use or inserted.