Add `files` component and remove `yaml` component
The `files` component allows a helper to register a key that will be parsed as filenames. Folder references will be fully expanded. If the helper defines a `callback stack` with the name of a stack then the `LoadRegisteredKeyFile` message will be sent to that stack for each file reference when the application is being loaded. The helper can handle this message in order to act on the file. For example, the helper might parse a yaml file.

The `yaml` component was removed in favor of the more generic `files` component.