Add "files" registered key component  (#40)
Helpers can register keys that should be processed but the framework. The framework already supported a "ui" type and a "yaml" type. "yaml" is being removed with this PR and replaced with a more generic "files" type. The framework will treat the registered key as a list of file references that will be processed when reading in the app.yml file (folder refs will be expanded and filenames completed).

When packaging an application the files will be copied into the package unless `distribute` is set to false. `distribute` can be set to "false" in the helper.yml file so that files for the key are never distributed. Or the developer can set `distribute` to false in the `app.yml` file for individual files.

Along with this commit is a `font_loader` helper which makes use of the new functionality. Here is an example from `helper.yml` file:

```
register components:
  1:
    key: fonts
    kind: files
```

If the helper needs to provide any special processing of files in the list then a `callback stack` property can be added to the entry in `register components`. The value is the name of a stack in memory. If set then a `LoadRegisteredKeyFile` message will be sent to the stack when each file in the key is loaded. One parameter is passed in which is the array for the file entry. The helper can modify the properties stored for that file in the Levure app config file. Here is an example that reads in a YAML file and stores the resulting array in the array that is passed in (`xFileA`). This updates the configuration information stored for your application.

```
register components:
  1:
    key: my-key
    kind: files
    callback stack: My Library
```

```
# "My Library" script
command LoadRegisteredKeyFile @xFileA
  put yamlFileToArray(xFileA["filename"]) into xFileA["data"]
  if the result is not empty then
    return "unable to read layout yaml file" && xFileA["filename"] && ":" && the result for error
  else
    return empty for value
  end if
end LoadRegisteredKeyFile
```