Don't run packager in test environment
When running in test environment no feedback is provided and some errors are swallowed up by the standalone builder. The upside to this change is that you see feedback during the build process and all errors are reported. The downside is that you have to press an OK button after each platform finishes building.