Add `postUIAssetPackaged` and `finalizePackagedAssets` messages to packager
`finalizePackagedAssets` is sent right before packaging up the assets for a platform. This message can be used to modify the app array prior to packaging for a specific platform. For example, it can be used to add to the list of behaviors that will be packaged.

`postUIAssetPackaged` is sent if the ui registered component has specified a `callback stackfile` in the helper.yml file. This message can be used to customize the output folder of the `ui` component.