Adds `build profiles filter` property
The `build profiles filter` is a comma delimited list of build profile names that affects packaging. It will include or exclude files based on the build profile being used to package the application. When the property is used with an asset (e.g. a helper or a file in the `copy files` section) the asset will be included in any of the build profiles in the list. If a build profile is prefixed with the `!` character then the asset will be included in any build profile except that one.