Dispatch all mementos with messages and fix cleanup
Previously a callback message and callback target could be associated with an undo type. These properties would be used when storing a memento but not when restoring mementos. Restoring send one memento at a time whereas restoring sent all mementos associated with an undo entry at once.

It has been decided that all mementos should be sent for storing/restoring/cleaning up as the undo implementation may need access to all of them to handle undo propertly. Given that change `pRemoveFromStackCallback` nd `pCallbackTarget` have been removed as parameters for `undoRegisterType`.

`undoCleanupMementos` is now sent when purging an undo/redo entry from the queue. This will be triggered when calling `undoReset` or `undoCleanupStack`