Add concept of `template` ui components
LiveCode does not allow controls on a password protected stack to be copied. This commit adds a new `templates` key to `app.yml` for `ui` components that should not have their binary stacks encrypted. Any script only stacks will be encrypted. If your application uses lots of templates for Data Grid's or other purposes then this makes your `app.yml` file cleaner as you don't have to explicity exclude multiple binary stacks from being password protected.

This commit also includes a fix when encrypting ui folders with multiple ui stacks. If a `ui` component folder had more than one ui stack in it only the encryption settings for the first stack would be honored. Now the packager generates the list of stacks that are not encrypted prior to moving the stacks into the distribution folder so that all stack encryption settings are honored.