Add NSUserDefaults extension for preferences on macOS and iOS
In LC >= 9 macOS and iOS now use an extension that wraps the NSUserDefaults APIs.

As part of the update an `engine version` key was added for externals/extensions and a `platform`
key was added for extensions. This allows the extension to be used on `macos,ios` in LC

engine_version setting allows you to specify a major.minor version of the engine that the external/extension supports. For example: <9.0 or >=9.0. Example usage is the preferences helper. The old macOS external is to be used before LC 9. The extension should be used in LC `>=9.0`.

In addition, a Levure bug where prefs for both "user" and "shared" were always being saved when the app shuts
down has been fixed. This would throw an error although the error would not be reported unless you had the debugger
running on iOS. A `prefsIsCategoryConfigured()` handler has been added so that Levure can check if prefs should
be saved for a particular category.

The preferences helper is now smarter about how it stores arrays and binary data on macOS.
`is strictly a binary string` is used to determine if a value shoul be stored as binary or
as a string.