Add basic `engine version` condition for externals/extensions
engine_version setting allows you to specify a major.minor version of the engine that the external/extension supports. For example: <9.0 or >=9.0. Example usage is the preferences helper. The old macOS external is to be used before LC 9. The extension should be used in 9.0 and later.

In addition, extensions can now have a `platform` property that lists the supported platforms for the extension.

Extensions and externals that don't match the engine version or platform will not be distributed when packaging the application.