Allow standalone builder to report errorrs to user
If the standalone builder encounters an error in test mode it doesn't return that error as `the result` of revSaveAsStandalone. The error is only acknowledged through an answer dialog.

This commit removes `revSetTestEnvironment true` allowing the standalone builder to report the error to the user. Levure logs that error but doesn't display it a second time to the user.