Add `wait with messages` in order to workaround IDE reloading wrong stacks
When building a standalone Levure makes a copy of the standalone stack that is passed off to the LiveCode standalone builder. Previously Levure would packaging in testing mode so that the IDE was silent. This suppressed important error messages that were not returned to the `revSaveAsStandalone` caller.

Now that testing mode is no longer being used the IDE throws up alerts about stacks with the same name being loaded. Levure is very careful about deleting the standalone stack and the temporary versions so this error shouldn’t be popping up. My best guess is that an IDE component is being sent a message during an event that occurs during the standalone building process and is reloading the stack the code just removed from memory behind our back. Adding `wait 1 second with messages` solves the problem. Not my favorite solution not worth spending any more time on either.