Add levureSetApplicationFont and levureSetApplicationFontSize
The font settings of the stack that the standalone is built with affect the whole application. When in the IDE that stack is `home`. In the standalone it is the stack that the Levure behavior is assigned to. These handlers will set the font of the correct stack so that you can see what your application will look like in the IDE.