Set mapping for libraries in extension’s code folder
An extenion can use a third party library (.dylib, .so, .dll, .framework, etc.). If it does then the library is placed in subfolder of a ./code folder for development and in the ./Externals folder when packaged. This update supports automatic loading of libraries. This is primarily used in helpers.

An extension definition can now include a `code` key which is a list. Each element is the name of the library to load, with or without the extension. If the extension is not provided then it will be guessed depending on the platform the applciation is running on. Using the library name, laodExtensionCode will figure out the correct path and set the revLibraryMapping[libraryname] property to the full path to the library. This is what allows FFI to resolve bindings in the LCB code.

Examples from helper.yml fles. This example tells Levure to map a libhunspell.dll file to the libhunspell library.

```
extensions:
  - filename: hunspell.lcm
    source: hunspell.lcb
    code:
      - libhunspell.dll
    platform: windows
```

This example will map libtidy.dylib, libtidy.so, or libtidy.dll to the libtidy library depending on which platform the app is running on.

```
extensions:
  - filename: htmltidy.lcm
    source: htmltidy.lcb
    code:
      - libtidy
```