Always create a .pkg when building for macOS
W/hen building for the 'mac app store'  build profile a .pkg is created to upload to the Mac App Store. Since a .pkg file can also be used by device management systems such as JAMF, a .pkg file is now generated when building for macOS in general.

Note that different signing certificates are used for the MAS vs a .pkg that is distributed outside of the MAS.