Update helper callback stack paths in packaged app
This patch resolves an error where the registerd key file callback stacks paths
are retained from the build and not resolved at runtime for the packaged app.
An error is therefore thrown in the packaged app at startup if there is both a
callback stackfile and distributed files.

An error will still be thrown if the callback stackfile was not distributed,
however, that is a valid error given the registered key files were distributed
and they require the callback stackfile.