Fix Windows build and packaging default folder issues (#218)
* Fix reload stack issue when building for Windows x86-64

This patch fixes an issue with reloading the framework stack when building for
windows as a previous patch introduced an issue where Windows x86 was always
built with Windows x86-64 rather than separately.

* Ensure default folder is reset after packaging

This patch updates the packager to make sure the default folder is reset when
packaging is complete. This ensures the default folder pointing to a deleted
folder does not cause a packaging failure on the second packaging in a session.